requirements.txt
flask
flake8
pandas
geopandas
xlrd
xlsxwriter
python-dotenv
googlemaps
flask-sqlalchemy
flask-login
psycopg2-binary
flask-wtf
email_validator
dash
git+https://github.com/dvrpc/tmc-summarizer.git